api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
